refactor(KERN): remove legacy design FilesUpload
