+++ app/components/formElements/filesUpload/FilesUploadHeader.tsx
@@ -9,8 +9,8 @@ export const FilesUploadHeader = ({
 }: FilesUploadHeaderProps) => {
   return (
     <div>
-      <p className="ds-label-02-bold font-black pb-6">{title}</p>
-      <p className="ds-body-01-reg text-gray-800">{description}</p>
+      {title && <p className="kern-label pb-6">{title}</p>}
+      {description && <p className="kern-hint">{description}</p>}
     </div>
   );
 };
